refactor: web search func into web search class method
